doc_ai_system/requirements.txt
google-genai
fastapi
uvicorn
python-dotenv
pydantic
python-multipart
